pyproject.toml
[tool.poetry]
name = "prototype"
version = "0.1.0"
description = ""
authors = ["Yonghyuk Yoo <yonghyuk_yoo@tmax.co.kr>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.3.0"
numpy = "^1.21.0"
requests = "^2.25.1"
aiohttp = "^3.7.4"
Flask-Caching = "^1.10.1"
statsmodels = "^0.12.2"
sklearn = "^0.0"
fastapi = "^0.66.0"
httpx = "^0.18.2"
fastapi-cache2 = "^0.1.4"
influxdb-client = "^1.19.0"
pystan = "2.19.1.1"
fbprophet = "^0.7.1"
dash = "^1.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
